Create Hard_Question_2
Hard Question #2
Write a function called longest_morse_code_words which takes in a list of strings. The output of longest_morse_code_words should be a list of the strings that were passed in, but ordered by the length of their Morse Code equivalent in descending order. If the length of Morse Code is equal, order the words alphabetically. For convenience, the full table for the 26 letters of the English alphabet is given below:
[".-","-...","-.-.","-..",".","..-.","--.","....","..",".---","-.-",".-..","--","-.","---",".--.","--.-",".-.","...","-","..-","...-",".--","-..-","-.--","--.."]



Here's an example:
words = ["gin", "zen", "gig", "msg"]
longest_morse_code_words(words) 

# The Morse Code equivalent of the words above is ["--...-.", "--...-.", "--...--.", "--...--."]. Ordering by length, it's [7, 7, 8, 8]. When we alphabetize based off of length in descending order, the output should be
# ['gig', 'msg', 'gin', 'zen']