Rename Hard_Question_1_letters_in_string to Hard_Question_1_Letters_in_string
Hard Question #1
Write a function called number_of_occurrences that takes in a string and outputs the the number of occurrences that a character shows up in a dictionary. This dictionary should be ordered by the number of occurrences in descending order. You can ignore spacing, capitalization, and punctuation.

Examples:
number_of_occurrences('aaabbc')
# returns {'a': 3, 'b': 2, 'c': 1}
number_of_occurrences('An apple a day. Will keep the doctor away!')
# returns {'a': 6, 'e': 4, 'p': 3, 'l': 3, 'y': 2, 'w': 2, 't': 2, 'd': 2, 'o': 2, 'h': 1, 'n': 1, 'c': 1, 'i': 1, 'k': 1, 'r': 1}