Create Medium_Question_4_Add_Adjacent_Numbers
Medium Question #4
Write a function called add_neighbors that takes a list of numbers and adds each number in the list to the number next to it. The list will always have an even number of numbers in it. In the end you should return a list of neighbors added up.

Example:
add_neighbors([1, 2, 3, 4, 5, 6]) # should return [3, 7, 11]
add_neighbors([12, 32, 35, 28, 10, 18, 55, 781]) # should return [44, 63, 28, 836]